Move chat checks to dedicated component
